# openapi.yaml
openapi: 3.0.3
info:
  title: Unico Napp API
  version: 1.0.0
  description: |
    Documentação OpenAPI da API Unico Napp.
    Cada recurso contém explicações e exemplos de requisição/ resposta.
servers:
  - url: http://localhost:5000
    description: Servidor local de desenvolvimento
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SimpleErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid API key
    ErrorItem:
      type: object
      properties:
        codigo_erro:
          type: integer
          example: 1001
        descricao:
          type: string
          example: username - Possui valor inesperado ou vazio
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
  responses:
...
    SellerNotFound:
      description: Seller was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - codigo_erro: 1014
                descricao: vendedor não encontrado
...
    SellerValidationError:
      description: Seller validation failed
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/SimpleErrorResponse'
              - $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalid_cnpj:
              value:
                errors:
                  - codigo_erro: 1006
                    descricao: CNPJ inválido
            duplicate_cnpj:
              value:
                errors:
                  - codigo_erro: 1012
                    descricao: CNPJ já cadastrado
...
paths:
...
  /api/v1/orders/seller/{seller_id}:
    get:
      summary: Pedidos por vendedor
      description: Retorna todos os pedidos associados a um vendedor.
      parameters:
        - name: seller_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Pedidos do vendedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
...
  /api/v1/sellers:
    get:
      summary: Listar vendedores
      description: Retorna fornecedores cadastrados com paginação.
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de vendedores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Seller'
    post:
      summary: Criar vendedor
      description: Cadastra um novo vendedor utilizando CNPJ válido.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SellerCreateRequest'
      responses:
        '201':
          description: Vendedor criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  cnpj:
                    type: string
                    example: 12345678000123
                  nome:
                    type: string
                    example: João Silva
                  message:
                    type: string
                    example: created
        '400':
          $ref: '#/components/responses/SellerValidationError'
        '409':
          description: Seller already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                errors:
                  - codigo_erro: 1012
                    descricao: CNPJ já cadastrado
  /api/v1/sellers/{id}:
    get:
      summary: Consultar vendedor por ID
      description: Retorna os dados de um vendedor específico.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Vendedor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Seller'
        '404':
          $ref: '#/components/responses/SellerNotFound'
...